+++ src/cjdk/__init__.py
@@ -12,15 +12,25 @@ from ._api import (
     list_jdks,
     list_vendors,
 )
+from ._exceptions import (
+    CjdkError,
+    ConfigError,
+    InstallError,
+    JdkNotFoundError,
+)
 from ._version import __version__ as __version__
 
 __all__ = [
     "cache_file",
     "cache_jdk",
     "cache_package",
+    "CjdkError",
     "clear_cache",
+    "ConfigError",
+    "InstallError",
     "java_env",
     "java_home",
+    "JdkNotFoundError",
     "list_jdks",
     "list_vendors",
 ]